[MAT-391] Outline platform_code()
