+++ cmake/buildutils.py
@@ -1,0 +1,18 @@
+#
+# Copyright (C) 2014 - present by OpenGamma Inc. and the OpenGamma group of companies
+#
+# Please see distribution for license.
+#
+
+import platform
+
+def platform_code():
+    """Returns the current platform code, following the usual convention
+    for OG-Maths."""
+    p = platform.system()
+    if p == 'Linux':
+        return 'lnx'
+    elif p == 'Darwin':
+        return 'osx'
+    else:
+        return 'win'
